services/knowledge-base/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
asyncpg
alembic
jinja2
python-multipart
python-docx
pypdf
psycopg2-binary
